// src/app/Models/Clientes/Juridicos/PatrimonioModel.ts
export class PatrimonioModel {
    IdJuridicoPatrimonio: number = 0;
    ActivosCorrientes: number = 0;
    ActivosNoCorrientes: number = 0;
    OtrosActivos: number = 0;
    ObligacionesFinancieras: number = 0;
    CuentasPorPagar: number = 0;
    OtrosPasivos: number = 0;
    IdTercero: number = 0;
}